+++ src/icons/logos/YouTube.tsx
@@ -27,6 +27,7 @@ const YouTube = ({
                 strokeMiterlimit: 2
             }}
             viewBox="0 0 33 23"
+            height={height}
         >
             <path
                 d="M724 1425.29h31.271v22.253H724z"